API Changes

+++ src/foliage/foliage.cpp
@@ -14,7 +14,7 @@ class RNG {
 	public:
 	RNG(unsigned seed=0) : m_seed(seed) {}
 	unsigned rand() { m_seed = m_seed * 1103515245 + 12345; return m_seed&0x7fffffff; }
-	float    randf() { return (float)rand() / 0x7fffffff; }
+	float    randf() { return (float)rand() / (float)0x7fffffff; }
 	float    randf(const Rangef& r) { return r.min + randf() * r.size(); }
 };
